updated simulation and parm_dicts
removed make_param_dicts.py and implemented the functions with the simulation class. Makes more sense to keep the functions combined with the class alongside modify_inp and modify_mol. Might suggest creating a class attribute inp_path and mol_path so that users won't have to specify a path when calling the function.